+++ index.js
@@ -42,6 +42,7 @@ const safeObject = require('./util/safe_object');
  * @property {false} [caniuse] Disable CanIUse embeds.
  * @property {false} [youtube] Disable YouTube embeds.
  * @property {false} [wistia] Disable Wistia embeds.
+ * @property {false} [vimeo] Disable Vimeo embeds.
  * @property {false} [twitter] Disable Twitter embeds.
  * @property {false} [instagram] Disable Instagram embeds.
  * @property {false} [underline] Disable underline syntax.
@@ -139,6 +140,10 @@ module.exports = (md, options) => {
 
     if (optsObj.wistia !== false) {
         md.use(require('./rules/embeds/wistia'), safeObject(optsObj.wistia));
+    }
+
+    if (optsObj.vimeo !== false) {
+        md.use(require('./rules/embeds/vimeo'), safeObject(optsObj.vimeo));
     }
 
     if (optsObj.twitter !== false) {