+++ CHANGELOG.md
@@ -18,6 +18,8 @@ See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
 - (minor) Add Instagram embeds
+- (minor) Add Vimeo embeds
+
 
 ## v1.5.1 - 2f1f346